+++ src/components/domain/GroupInfoModal/index.jsx
@@ -1,12 +1,18 @@
-import { Button, Icon, Modal, SearchBar, Tag, Text } from '@/components/base';
+import { Button, Icon, SearchBar, Tag, Text } from '@/components/base';
 import { Introduction, Member, MemberList } from '@/components/domain/groupInfo';
 import { useAuthContext } from '@/context/AuthProvider';
 import { useGroupContext } from '@/context/GroupProvider';
 import { useUserContext } from '@/context/UserProvider';
 import useInput from '@/hooks/useInput';
 import { COLOR } from '@/styles/color';
-import styled from '@emotion/styled';
 import { useEffect, useState } from 'react';
+import {
+  StyledButtonWrapper,
+  StyledContentContainer,
+  StyledHeaderContainer,
+  StyledMemberListContainerLabel,
+  StyledModal,
+} from './styles';
 
 function GroupInfoModal({ group, visible, onClose, ...props }) {
   const {
@@ -131,48 +137,3 @@ function GroupInfoModal({ group, visible, onClose, ...props }) {
 }
 
 export default GroupInfoModal;
-
-const StyledModal = styled(Modal)`
-  width: 70rem;
-  @media (max-width: 624px) {
-    width: 100%;
-  }
-`;
-
-const StyledHeaderContainer = styled.div`
-  padding: 0.5rem 2rem 0 2rem;
-  margin-bottom: 3rem;
-  display: flex;
-  justify-content: space-between;
-
-  & > :nth-of-type(2) {
-    width: 80%;
-    padding-top: 2rem;
-  }
-
-  & > span {
-    text-align: center;
-  }
-`;
-
-const StyledContentContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  padding: 2rem 5%;
-`;
-
-const StyledMemberListContainerLabel = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  margin: 3rem 0 1rem 0;
-`;
-
-const StyledButtonWrapper = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  width: 100%;
-  margin-top: 2rem;
-`;
